requirements.txt
torch>=1.10.0,<2.0.0
torchvision>=0.11.1,<1.0.0
torchaudio>=0.10.0,<1.0.0
transformers>=4.20.0
datasets>=2.0.0
evaluate>=0.3.0
wandb>=0.12.0
PyYAML>=6.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
accelerate>=0.12.0 # Required by Trainer
segmentation-models-pytorch>=0.3.0 # Optional: if you want other models like Unet easily
tqdm>=4.60.0
Pillow>=9.0.0